clean up in debug façade, moved important functions in mainStateFacade
