+++ gameModel/gamemodel.json
@@ -4,7 +4,7 @@
   "comments" : null,
   "createdById" : null,
   "createdByName" : null,
-  "createdTime" : 1727430177202,
+  "createdTime" : 1729504644346,
   "description" : null,
   "items" : [ {
     "@class" : "ListDescriptor",
@@ -6228,6 +6228,15 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
+    "contentKey" : "UIfacade/mainStateFacade",
+    "contentType" : "application/typescript",
+    "libraryType" : "ClientScript",
+    "refId" : "GameModelContent:1220808:lhMLFW",
+    "version" : null,
+    "visibility" : "PRIVATE"
+  }, {
+    "@class" : "GameModelContent",
+    "content" : "",
     "contentKey" : "UIfacade/methaneFacade",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
@@ -6703,7 +6712,7 @@
     "version" : null,
     "visibility" : "PRIVATE"
   } ],
-  "name" : "MIMMS main sim",
+  "name" : "MIMMS - MIM-259 XGO",
   "pages" : { },
   "properties" : {
     "@class" : "GameModelProperties",
