+++ gameModel/libs/ClientScript/gameMap/layers/mainSimLayer.ts
@@ -7,10 +7,10 @@ import {
   SelectedPositionType,
   AvailablePositionType,
 } from '../../game/common/events/defineMapObjectEvent';
+import { getCurrentState } from '../../game/mainSimulationLogic';
 import { FeatureCollection } from '../../gameMap/types/featureTypes';
 import { getEmptyFeatureCollection } from '../../gameMap/utils/mapUtils';
 import { getTranslation } from '../../tools/translation';
-import { getCurrentState } from '../../UIfacade/debugFacade';
 
 /************
  *
